Changes to fixel helpers to support -force option
- check_fixel_directory() has been split into check_fixel_directory_in() and check_fixel_directory_out(), as the requisite behaviour for these two cases is quite different.
- Instead of the old check_fixel_directory(), which received two booleans "create_if_missing" and "check_if_empty", new function check_fixel_directory_out() receives two different booleans, which state whether or not the output fixel directory is going to have new index and directions files written; these control whether or not the contents of a pre-existing non-empty output directory should be wiped (only if a new index image is to be created), and whether or not it should be permissible for the output directory location to not yet exist.
- Moved function definitions to .cpp file, and removed FORCE_INLINE directive, as these functions are not performance-critical.
- Change some fixel command tests so that files / directories are cleaned up by the default cleanup mechanisms.
