+++ cmd/fixelfilter.cpp
@@ -15,6 +15,8 @@
  */
 
 
+#include <set>
+
 #include "command.h"
 #include "header.h"
 #include "image.h"
