+++ cmd/peaks2fixel.cpp
@@ -93,7 +93,7 @@ void run ()
     INFO ("Peaks have variable amplitudes; will create additional fixel data file \"" + dataname + "\"");
   }
 
-  Fixel::check_fixel_directory (argument[1], true, true);
+  Fixel::check_fixel_directory_out (argument[1], true, true);
 
   // Easiest if we first make the index image
   const std::string index_path = Path::join (argument[1], "index.mif");
