+++ cmd/voxel2fixel.cpp
@@ -51,7 +51,7 @@ void run ()
 {
   auto scalar = Image<float>::open (argument[0]);
   std::string input_fixel_directory = argument[1];
-  Fixel::check_fixel_directory (input_fixel_directory);
+  Fixel::check_fixel_directory_in (input_fixel_directory);
   auto input_fixel_index = Fixel::find_index_header (input_fixel_directory).get_image<index_type>();
   check_dimensions (scalar, input_fixel_index, 0, 3);
