+++ docs/reference/commands/fixelreorient.rst
@@ -16,13 +16,15 @@ Usage
     fixelreorient [ options ]  fixel_in warp fixel_out
 
 -  *fixel_in*: the input fixel directory
--  *warp*: a 4D deformation field used to perform reorientation. Reorientation is performed by applying the Jacobian affine transform in each voxel in the warp, then re-normalising the vector representing the fixel direction
--  *fixel_out*: the output fixel directory. If the the input and output directories are the same, the existing directions file will be replaced (providing the -force option is supplied). If a new directory is supplied then the fixel directions and all other fixel data will be copied to the new directory.
+-  *warp*: a 4D deformation field used to perform reorientation
+-  *fixel_out*: the output fixel directory
 
 Description
 -----------
 
 Reorientation is performed by transforming the vector representing the fixel direction with the Jacobian (local affine transform) computed at each voxel in the warp, then re-normalising the vector.
+
+If the the input and output directories are the same, the existing directions file will be replaced (providing the -force option is supplied). If a new directory is supplied then the fixel directions and all other fixel data will be copied to the new directory.
 
 Options
 -------
