+++ run_tests
@@ -231,6 +231,7 @@ if [[ ! -z $testsdir && ! -z $datadir ]]; then
   echo -n "Cleaning up temporary files... "
   rm -f $datadir/tmp*.*
   rm -rf $datadir/tmp-*/
+  rm -rf $datadir/tmp_*/
   rm -rf $datadir/*-tmp-*/
   echo OK
 fi
