+++ testing/binaries/tests/fixelreorient
@@ -1,1 +1,1 @@
-fixelreorient fixel_image fod_warp.mif fixelreorient_tmp -force && testing_diff_fixel fixelreorient_tmp fixelreorient -frac 0.001 && rm -rf fixelreorient_tmp
+fixelreorient fixel_image fod_warp.mif fixelreorient_tmp -force && testing_diff_fixel fixelreorient_tmp fixelreorient -frac 0.001
